refactor sm, fix policy conf to int16, tested with rf-sim

+++ examples/xApp/go/FlexPolicy/non_rt_emu/max_prb_utli/non_rt_max_prb_utli_main.go
@@ -24,7 +24,7 @@ var server1URL = "http://127.0.0.1:7000/api/policy"
 func main() {
 	rand.Seed(time.Now().UnixNano())
 
-	prbs := []int{100, 80, 20, 90, 20, 90, 20, 90, 20, 90}
+	prbs := []int16{100, 80, 20, 90, 20, 90, 20, 90, 20, 90}
 
 	for _, maxPrbUtil := range prbs {
